Move `mangled()` under `generate` package
